# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.APIResponse:
    properties:
      data: {}
      error:
        type: string
      success:
        type: boolean
    type: object
...
  types.DepartmentLookup:
    properties:
      id:
        type: integer
      manager:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@kartezya.com
    name: API Support
  description: |-
    This is a comprehensive HR Management System API built with Go and Gin framework.
    It provides endpoints for managing employees, leave requests, companies, departments, and more.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Kartezya HR Management System API
  version: "1.0"
paths:
...
  /lookup/departments-by-company:
    get:
      description: Get departments filtered by company ID as lookup data (public API)
      parameters:
      - description: Company ID
        in: query
        name: company_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/types.DepartmentLookup'
                  type: array
              type: object
      summary: Get departments by company for lookup
      tags:
      - Lookup
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"